requirements.txt
uwaterlooapi
requests
python-telegram-bot
python-twitter
